Assign charmverse roles from discord roles
