+++ pages/api/discord/importRoles.ts
@@ -4,9 +4,9 @@ import { prisma } from 'db';
 import { onError, onNoMatch, requireSpaceMembership, requireUser } from 'lib/middleware';
 import nc from 'next-connect';
 import { withSessionRoute } from 'lib/session/withSession';
-import { Role } from '@prisma/client';
 import { handleDiscordResponse } from 'lib/discord/handleDiscordResponse';
 import { createRolesFromDiscord } from 'lib/role/createRolesFromDiscord';
+import { assignRolesFromDiscord } from 'lib/role/assignRolesFromDiscord';
 import { DiscordUser } from './connect';
 
 const handler = nc({
@@ -158,90 +158,14 @@ async function importRoles (req: NextApiRequest, res: NextApiResponse<ImportRole
     });
   });
 
-  for (const discordGuildMember of discordGuildMembers) {
-    const unassignedRoles: string[] = [];
-    try {
-      // No need to add role for a bot
-      if (discordGuildMember.user && !discordGuildMember.user.bot) {
-        // Find the charmverse user whose discord id matches with the guild member
-        const charmverseUser = await prisma.user.findFirst({
-          where: {
-            discordUser: {
-              discordId: discordGuildMember.user.id
-            }
-          }
-        });
+  const failedRoleAssignments = await assignRolesFromDiscord(rolesRecord, discordGuildMembers, spaceId);
 
-        // Some user might not have connected with discord
-        if (charmverseUser) {
-          // Check if the user has a role in the current space, this is to ensure that the user belongs to this space
-          const charmverseUserSpaceRole = await prisma.spaceRole.findFirst({
-            where: {
-              spaceId,
-              userId: charmverseUser.id
-            }
-          });
-
-          // If the user is a part of the workspace assign the role to that user
-          if (charmverseUserSpaceRole) {
-            // Loop through all the roles the discord user has
-            for (const role of discordGuildMember.roles) {
-              const roleInRecord = rolesRecord[role];
-              try {
-                // If the role was created/fetched successfully
-                if (roleInRecord?.charmverse) {
-                  // Check if the user already has the same role
-                  const charmverseUserExistingRole = await prisma.spaceRoleToRole.findFirst({
-                    where: {
-                      roleId: roleInRecord.charmverse.id,
-                      spaceRoleId: charmverseUserSpaceRole.id!
-                    }
-                  });
-                    // Only assign the role if the role wasn't already assigned to the charmverse user
-                  if (!charmverseUserExistingRole) {
-                    await prisma.spaceRoleToRole.create({
-                      data: {
-                        role: {
-                          connect: {
-                            id: roleInRecord.charmverse.id
-                          }
-                        },
-                        spaceRole: {
-                          connect: {
-                            id: charmverseUserSpaceRole.id!
-                          }
-                        }
-                      }
-                    });
-                  }
-                }
-                else {
-                  // We couldn't create the role so it wont be assigned
-                  unassignedRoles.push(roleInRecord?.discord.name);
-                }
-              }
-              catch (_) {
-                unassignedRoles.push(roleInRecord?.discord.name);
-              }
-            }
-          }
-        }
-      }
-    }
-    catch (_) {
-      // This block will be reached if
-      // 1. User doesn't have a charmverse account connected with discord
-      // 2. User is not part of the charmverse workspace
-    }
-
-    if (unassignedRoles.length !== 0) {
-      importErrors.push({
-        action: 'assign',
-        username: discordGuildMember.user?.username ?? 'N/A',
-        roles: unassignedRoles
-      });
-    }
-  }
+  failedRoleAssignments.forEach(failedRoleAssignment => {
+    importErrors.push({
+      action: 'assign',
+      ...failedRoleAssignment
+    });
+  });
 
   res.status(200).json({
     error: importErrors
